Rename amazonimport to zipimport, ahead of it serving any zip
The package stages a zip, hashes every entry, spots duplicates against the
library and imports what is left as pending documents. None of that is about
Amazon; only the entry filter is. A generic "import a zip of my documents"
flow wants the rest of it verbatim, so the package is about to grow a source
argument and serve both. Rename it first, on its own, so the commit that
changes behaviour is small enough to read.

Pure rename: the package clause in all seven files, the import path and
qualifier in appapi and appwire, and one comment in perfilecaps that named the
package. No symbol, signature or route changed; the existing tests pass
untouched.
