+++ backend/internal/zipimport/import.go
@@ -1,4 +1,4 @@
-package amazonimport
+package zipimport
 
 import (
 	"archive/zip"
