+++ backend/internal/zipimport/staging.go
@@ -1,4 +1,4 @@
-package amazonimport
+package zipimport
 
 import (
 	"archive/zip"
