+++ backend/internal/zipimport/staging_test.go
@@ -1,4 +1,4 @@
-package amazonimport
+package zipimport
 
 import (
 	"bytes"
